Move logic for adding/removing line items from user to cart.
