+++ app/models/user.rb
@@ -11,16 +11,4 @@ class User < ActiveRecord::Base
   def create_cart
     Cart.create(user: self)
   end
-
-  def add_to_cart(book)
-    cart.line_items.create!(book_id: book.id, quantity: 1)
-  end
-
-  def remove_from_cart(book)
-    cart.line_items.find_by_book_id(book.id).destroy
-  end
-
-  def in_cart?(book)
-    cart.line_items.any? { |line_item| line_item[:book_id] == book.id }
-  end
 end
